Update makefile to have one go build rule
This works for both build/soarca and build/soarca-conversion, and
properly depends on all go files to detect rebuilds. The default build
target always triggers this, as swag will touch api/docs.go; filtering
this out broke the dependency for reasons beyond comprehension, but
since builds are quite quick this is okay for now.